+++ mne/time_frequency/tests/test_tfr.py
@@ -95,6 +95,20 @@ def test_time_frequency():
     assert_array_almost_equal(power.data, power_picks_avg.data)
     assert_array_almost_equal(itc.data, itc_picks.data)
     assert_array_almost_equal(power.data, power_evoked.data)
+    # complex output
+    assert_raises(ValueError, tfr_morlet, epochs, freqs, n_cycles,
+                  return_itc=False, average=True, output="complex")
+    assert_raises(ValueError, tfr_morlet, epochs, freqs, n_cycles,
+                  output="complex", average=False, return_itc=True)
+    epochs_power_complex = tfr_morlet(epochs, freqs, n_cycles,
+                                      output="complex", average=False,
+                                      return_itc=False)
+    epochs_power_2 = abs(epochs_power_complex)
+    epochs_power_3 = epochs_power_2.copy()
+    epochs_power_3.data[:] = np.inf  # test that it's actually copied
+    assert_array_almost_equal(epochs_power_2.data, epochs_power_picks.data)
+    power_2 = epochs_power_2.average()
+    assert_array_almost_equal(power_2.data, power.data)
 
     print(itc)  # test repr
     print(itc.ch_names)  # test property
@@ -510,11 +524,11 @@ def test_compute_tfr():
         # Check exception
         if (func == tfr_array_multitaper) and (output == 'phase'):
             assert_raises(NotImplementedError, func, data, sfreq=sfreq,
-                          frequencies=freqs, output=output)
+                          freqs=freqs, output=output)
             continue
 
         # Check runs
-        out = func(data, sfreq=sfreq, frequencies=freqs, use_fft=use_fft,
+        out = func(data, sfreq=sfreq, freqs=freqs, use_fft=use_fft,
                    zero_mean=zero_mean, n_cycles=2., output=output)
         # Check shapes
         shape = np.r_[data.shape[:2], len(freqs), data.shape[2]]